Fix cannot create frame session bug
There were missing locks for execution contexts. This fix adds a
lock-based management for locks and uses an internal goroutine
to track whether an execution context is ready to use.

For now, it uses an infinite for loop with sleep. This is not
effective. So it should be fixed with something better.

There are some TODOs in the code that should be investigated.
Also the tests are missing from this commit. So they should
be added as well.

Fixes #125